[ts sdk] Remove deprecated types directory
