+++ sdk/typescript/src/builder/serializer.ts
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { SuiJsonValue, SuiMoveNormalizedType } from '../client/index.js';
-import { extractStructTag } from '../types/index.js';
 import { MOVE_STDLIB_ADDRESS, SUI_FRAMEWORK_ADDRESS } from '../utils/index.js';
 import { isValidSuiAddress } from '../utils/sui-types.js';
+import { extractStructTag } from './utils.js';
 
 const OBJECT_MODULE_NAME = 'object';
 const ID_STRUCT_NAME = 'ID';
