--- sdk/typescript/src/types/index.ts
@@ -1,140 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	CheckpointedObjectId,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DisplayFieldsBackwardCompatibleResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	DisplayFieldsResponse,
-	/** @deprecated This type will be removed in a future version */
-	GetOwnedObjectsResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	MovePackageContent,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ObjectContentFields,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	ObjectRead,
-	/** @deprecated This type will be removed in a future version */
-	ObjectStatus,
-	/** @deprecated This type will be removed in a future version */
-	ObjectType,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type Order,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	PaginatedObjectsResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiGasData,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveObject,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMovePackage,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectData,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type SuiObjectDataFilter,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectDataOptions,
-	/** @deprecated This type will be removed in a future version */
-	type SuiObjectDataWithContent,
-	/** @deprecated This type will be removed in a future version */
-	SuiObjectInfo,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectRef,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiObjectResponseError,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	type SuiObjectResponseQuery,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiParsedData,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiRawData,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiRawMoveObject,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiRawMovePackage,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getMoveObject,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getMoveObjectType,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getMovePackageContent,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectDeletedResponse,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectDisplay,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectFields,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectId,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectNotExistsResponse,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectOwner,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectPreviousTransactionDigest,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectReference,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectType,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getObjectVersion,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getSharedObjectInitialVersion,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	getSuiObjectData,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	hasPublicTransfer,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	isImmutableObject,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	isSharedObject,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	isSuiObjectResponse,
-} from './objects.js';
-
-export {
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	MoveCallMetric,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	MoveCallMetrics,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveAbilitySet,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveFunctionArgType,
-	/* @deprecated Use SuiMoveFunctionArgType[] from `@mysten/sui-js/client` instead */
-	SuiMoveFunctionArgTypes,
-	/* @deprecated Use SuiMoveFunctionArgType[] from `@mysten/sui-js/client` instead */
-	type SuiMoveFunctionArgTypesResponse,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveModuleId,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedField,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedFunction,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedModule,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedModules,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedStruct,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedStructType,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedType,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveNormalizedTypeParameterType,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveStructTypeParameter,
-	/** @deprecated Import type from `@mysten/sui.js/client` instead */
-	SuiMoveVisibility,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	extractMutableReference,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	extractReference,
-	/** @deprecated This method will be removed in a future version of the SDK */
-	extractStructTag,
-} from './normalized.js';
