--- sdk/typescript/src/types/subscriptions.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-export type Unsubscribe = () => Promise<boolean>;
